+++ crates/store/src/backend/s3/mod.rs
@@ -69,7 +69,7 @@ impl S3Store {
             },
         };
         let credentials = Credentials::new(
-            config.access_key.as_deref(),
+            config.access_key.value().await?.as_deref(),
             config.secret_key.secret().await?.as_deref(),
             config.security_token.secret().await?.as_deref(),
             config.session_token.secret().await?.as_deref(),